Moved common types to another folder Cleaned code
