+++ src/common/types/row.ts
@@ -1,0 +1,7 @@
+export type rowType = {
+    latitude: string;
+    longitude: string;
+    name: string;
+    type: string;
+    id: string;
+};
