+++ src/common/types/table-props.ts
@@ -1,0 +1,5 @@
+import {SearchTypeEnum} from "../../utils/constants";
+
+export type tableProps = {
+    type: SearchTypeEnum;
+};
