+++ src/pages/data-table-page.test.tsx
@@ -1,13 +1,11 @@
 import React from 'react'
 import 'jest-styled-components';
-import StartPage from "./start-page";
 import {ThemeProvider} from "styled-components";
 import {baseTheme} from "../styled/themes/base-theme";
-import {RoutesEnum, SEARCH_TYPE_DISPLAY_MAP, SearchTypeEnum} from "../utils/constants";
+import {SearchTypeEnum} from "../utils/constants";
 import {fireEvent, render} from "@testing-library/react";
 import {screen} from '@testing-library/dom'
 import DataTablePage from "./data-table-page";
-import {useNavigate} from "react-router-dom";
 
 const mockedUsedNavigate = jest.fn();
 
@@ -31,11 +29,11 @@ describe("Start Page tests", () => {
     });
 
     test("Data table Page redirects on button click", () => {
-        const {container} = (render(
+        render(
             <ThemeProvider theme={baseTheme}>
                 <DataTablePage type={SearchTypeEnum.DRONE_RES}/>
             </ThemeProvider>
-        ));
+        );
         const button = screen.getByText('Search');
         fireEvent.click(button);
