+++ src/services/getPopulationDensity.ts
@@ -19,7 +19,7 @@ const getPopulationDensity = (queryParams: { latitude: string, longitude: string
                     lang: 'en',
                     returnGeometry: 'false',
                     tolerance: '0',
-                    geometry: `{'x': ${longitude}, 'y': ${latitude}}`,
+                    geometry: `{"x": ${longitude}, "y": ${latitude}}`,
                 }
             })
             .then(response => res(response.data))
